# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dega server API",
        "title": "Dega API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7789",
    "basePath": "/",
    "paths": {
...
        "/core/permissions/organisations": {
            "get": {
                "description": "Get all organisation permissions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organisation_Permissions"
                ],
                "summary": "Show all organisation permissions",
                "operationId": "get-all-org-permissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Query",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github.com_factly_dega-server_service_core_action_permissions_organisation.paging"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create organisation permission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organisation_Permissions"
                ],
                "summary": "Create organisation permission",
                "operationId": "add-org-permission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Permission Object",
                        "name": "Permission",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/organisation.organisationPermission"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.OrganisationPermission"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/core/requests/organisations": {
            "get": {
                "description": "Get all organisation permissions requests",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organisation_Permissions_Request"
                ],
                "summary": "Show all organisation permissions requests",
                "operationId": "get-all-org-permissions-requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github.com_factly_dega-server_service_core_action_request_organisation.paging"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create organisation permission request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organisation_Permissions_Request"
                ],
                "summary": "Create organisation permission request",
                "operationId": "add-org-permission-request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Request Object",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/organisation.organisationPermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.OrganisationPermissionRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/core/requests/spaces/{request_id}/reject": {
            "post": {
                "description": "reject space permission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Space_Permissions_Request"
                ],
                "summary": "reject space permission",
                "operationId": "reject-space-permission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "request_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/core/webhooks/{webhook_id}": {
            "get": {
                "description": "Get webhook by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "Show a webhook by id",
                "operationId": "get-webhook-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Webhook ID",
                        "name": "webhook_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Webhook"
                        }
                    }
                }
            },
            "put": {
                "description": "Update webhook by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "Update a webhook by id",
                "operationId": "update-webhook-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Webhook ID",
                        "name": "webhook_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Webhook Object",
                        "name": "Webhook",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/webhook.webhook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Webhook"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete webhook by id",
                "tags": [
                    "Webhooks"
                ],
                "summary": "Delete webhook by id",
                "operationId": "delete-webhook-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Webhook ID",
                        "name": "webhook_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github.com_factly_dega-server_service_core_action_permissions_organisation.paging": {
            "type": "object",
            "properties": {
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/organisation.orgWithPermissions"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "github.com_factly_dega-server_service_core_action_request_organisation.paging": {
            "type": "object",
            "properties": {
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OrganisationPermissionRequest"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "model.Event": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "event": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tags": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.Medium": {
            "type": "object",
            "properties": {
                "alt_text": {
                    "type": "string"
                },
                "caption": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dimensions": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "meta_fields": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "space_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "model.OrganisationPermission": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "spaces": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
        "model.OrganisationPermissionRequest": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "spaces": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.Webhook": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Event"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tags": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "organisation.orgWithPermissions": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "featured_medium_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "medium": {
                    "$ref": "#/definitions/model.Medium"
                },
                "permission": {
                    "$ref": "#/definitions/model.OrganisationPermission"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
        "organisation.organisationPermission": {
            "type": "object",
            "required": [
                "organisation_id"
            ],
            "properties": {
                "organisation_id": {
                    "type": "integer"
                },
                "spaces": {
                    "type": "integer"
                }
            }
        },
        "organisation.organisationPermissionRequest": {
            "type": "object",
            "required": [
                "organisation_id"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "spaces": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "webhook.webhook": {
            "type": "object",
            "required": [
                "event_ids",
                "url"
            ],
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "event_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}